Add Room database implementation + Add dependency injection
1) Room database implementation
Add new module "data"

Add new file "Converters.kt"

Delete file "DummyEstateGenerator.kt" and package "service"

Estate.kt :
	- Update property type for "id"

EstateDao.kt :
	- Add new method "updateEstateData()"

EstateData.kt :
	- Update property type for "idEstate"

Add new file "EstateWithPhotosAndInteriorDao.kt"

FragmentNewEstate.kt :
	- Update method "createEstate()"

InteriorDao.kt :
	- Add method "insertInteriorData()"
	- Add method "updateInteriorData()"

InteriorData.kt :
	- Update property type for "idInterior"

ListEstatesViewModel.kt :
	- Rename method "updateViewModel()" to "updateDatabase()"
	- Update method "createNewEstate()"
	- Add new method "restoreData()"
	- Add new method "insertEstateInDatabase()"
	- Add new method "insertPhotoInDatabase()"
	- Add new method "insertInteriorInDatabase()"
	- Add new method "updateEstateInDatabase()"
	- Add new method "updateInteriorInDatabase()"

MainActivity.kt :
	- Update method "onCreate()"
	- Add new method "accessDatabase()"

PhotoDao.kt :
	- Add new method "getPhotos()"

PhotoData.kt :
	- Rename properties

RealEstateManagerDatabase.kt :
	- Remove method "getDatabase()"

Add new file "RealEstateRepository.kt"

Add new file "RepositoryModule.kt"

2) Dependency injection
Add new file "BaseApplication.kt"

Add new file "DatabaseModule.kt"

ListEstatesViewModel.kt :
	- Update constructor with @Inject annotation

MainActivity.kt :
	- Defines class as AndroidEntryPoint

3) Correction code :
FragmentListEstate.kt :
	- Update method "addObserversToViewModels()"

+++ app/src/main/java/com/openclassrooms/realestatemanager/model/Estate.kt
@@ -1,9 +1,5 @@
 package com.openclassrooms.realestatemanager.model
 
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
 /**
  * Defines a real estate.
  * @param id : Id of the estate
@@ -21,7 +17,7 @@ import androidx.room.PrimaryKey
  * @param selected : Define if a estate is currently being consulted (true) or not (false)
  */
 data class Estate(
-    val id: Int,
+    val id: Long,
     var type: String,
     var district: String,
     var price: Int,
